refactor: rename controllers/ package to api/
Having both controllers.py and controllers/ in the same directory causes
Python to shadow controllers.py with the package on import. Renaming to
api/ is also a cleaner convention: controllers.py for py4web page routes,
api/ for REST API endpoints.
